chore: use uint192 for exitId in processExit
The 'processExit' function was designed with exitId of uint256. As
all the other places uses uint192 for exitId, there is no need to keep
uint256 and can make code consistent.

+++ plasma_framework/contracts/mocks/exits/payment/PaymentStandardExitableMock.sol
@@ -17,7 +17,7 @@ contract PaymentStandardExitableMock is PaymentStandardExitable {
     }
 
     /** override and calls processStandardExit for test */
-    function processExit(uint256 _exitId) external {
+    function processExit(uint192 _exitId) external {
         PaymentStandardExitable.processStandardExit(_exitId);
     }
