+++ plasma_framework/contracts/src/framework/models/ExitModel.sol
@@ -4,6 +4,6 @@ library ExitModel {
     struct Exit {
         address exitProcessor;
         uint256 exitableAt;
-        uint256 exitId; // This is for each exit game contract to design
+        uint192 exitId; // This is for each exit game contract to design
     }
 }
